requirements.txt
openai>=1.0.0
tiktoken>=0.7.0
pillow>=10.0.0
pyyaml>=6.0
typer>=0.12.3
rich>=13.7.1
matplotlib>=3.9.0
aiofiles>=23.2.1 
pandas>=2.2.2 
numpy>=1.26.4 